Added a few more tests, including initial tests for methods other than GET.  Refactored.  Fixed compiler warning in class HTTPRequestWithTextOnlyPostContent_F.

+++ tests/HTTPRequestTests.cpp
@@ -162,7 +162,7 @@ public:
 	~HTTPRequestWithTextOnlyPostContent_F() {
 	}
 
-	int m_len;
+	size_t m_len;
 	char* m_post_buffer;
 };
